api/requirements.txt
Werkzeug==0.16.1
flask==1.1.1
flask-restplus==0.13.0
flask-bcrypt==0.7.1
flask-sqlalchemy==2.4.1
flask-migrate==2.5.2
flask-script==2.0.6
SQLAlchemy==1.3.13
python-dotenv==0.10.3
bcrypt==3.1.7
jsonschema==3.2.0
PyMySQL==0.9.3
python-gitlab==2.0.1